Refactor : Refactoring the conversion screen, better extracion
